+++ prefetcher/gasp/svm4ap/global.hpp
@@ -2,12 +2,13 @@
 #include <stdio.h>
 #include<iostream>
 #include <vector>
+#include<memory>
 
 using namespace std;
 
-#include "svm.hpp"
-#include "svm_impls.hpp"
-#include "input_buffer.hpp"
-#include "input_buffer_impls.hpp"
-#include "dictionary.hpp"
-#include "dictionary_impls.hpp"
+#include "svm/svm.hpp"
+#include "svm/svm_impls.hpp"
+#include "input_buffer/input_buffer.hpp"
+#include "input_buffer/input_buffer_impls.hpp"
+#include "dictionary/dictionary.hpp"
+#include "dictionary/dictionary_impls.hpp"
