+++ prefetcher/gasp/svm4ap/input_buffer/input_buffer.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "global.hpp"
+#include  "svm4ap/global.hpp"
 template<typename InputBufferEntry>
 class InputBuffer{
     public:
