+++ prefetcher/gasp/svm4ap/input_buffer/input_buffer_impls.cpp
@@ -1,10 +1,12 @@
 #include "input_buffer_impls.hpp"
 
+template<typename InputBufferEntry>
 std::optional<InputBufferEntry> StandardInputBuffer<InputBufferEntry>::read(uint64_t ip){
     auto entry = this->table.check_hit(InputBufferEntry(ip));
     return entry;
 }
 
+template<typename InputBufferEntry>
 void StandardInputBuffer<InputBufferEntry>::write (InputBufferEntry entry){
     this->table.fill(entry);
 }
